+++ fidget/src/core/eval/test/mod.rs
@@ -1,0 +1,5 @@
+//! Test suites for each evaluator type
+pub mod float_slice;
+pub mod grad_slice;
+pub mod interval;
+pub mod point;
